refactor: 전역상태 관리로 리팩토링
