+++ src/app/components/editor/DefaultEditor.module.scss
@@ -22,8 +22,3 @@
 
   // TODO heading 등 정의 필요
 }
-
-.prompt-menu {
-  position: relative;
-  width: 200px;
-}
