+++ src/app/components/editor/PromptInput.module.scss
@@ -27,3 +27,8 @@
     outline: none;
   }
 }
+
+.select-menu {
+  position: relative;
+  width: 200px;
+}
